cell-entrance WorldEvents in World queue
When bodies are added to the world, they start off in the invalid
state. When we need the world queue to be come valid, we bulk-process
the invalid bodies, adding their upcoming events to the world queue,
which is a priority queue, implemented with a SkipList prioritized by
event time.
I got as far as adding the upcoming cell-entrance events, including
their cell ranges. test8 dumps demo data to the console for debugging.