Canvas layers - test12
It looks like it’s better to render a background canvas once, and slide
that whole absolutely-positioned canvas around, instead of redrawing
the background every frame.
But there’s a limit to how big the background can be before the game
gets very slow. Maybe the background can’t fit into the GPU as a
texture when it’s too big.