Fracas 2 DEATH BRUSH
Now time slows down when gnomes with line of sight to the player get
close to the player. It makes the game less reflex-y and more
strategic, and encourages the player to get close dodge through gnomes
instead of keeping a fearful distance.
Also tweaked colors and added some (unused) zoom controls.