test34 world units and dimensional analysis
Now the world is in meters, and only terrain cells that are overlapping
the screen get drawn. It used to be hard-coded and a bit overkill.
Adjusted the cursor code to respect the viewport size, so that if I add
zooming in and out, the cursor will adjust correctly.
Discovered I accidentally got touch-editing working, sort of.