Fixed BitGrid "Sudden Square" bug!
Sometimes the bitgrid would fill or clear a while cell at the wrong
time. I fixed that, but I’m not sure I understand what was wrong with
the original approach. Something to do with mixing bitwise operations
and addition on the same var, I think, causing the var to lose al but
32 bits of accuracy?