Added event-stack Trigger pubsubs, especially MultiTrigger
Now games can subscribe to trigger up/down events, and get called in
the event handler stack, which is needed for browser permission stuff
like pointerlock.
MultiTrigger listens to its children and publishes up/down events only
when the consolidated up/down value changes. It’s great!
This is all needed so the “pause” button can get events when the
animation loop is stopped, which is coming up next…