LayeredEventDistributor
Now mouse and touch and keyboard events can handled one layer at a
time, front to back in the UI. If one layer consumes an event, layers
behind that will not receive it. That means that the world trackball,
for instance, does not need to know about the layout of the triggers in
the UI.
LayeredEventDistributor can also create fake event target objects, one
per layer, for other bits of code to use for doing addListener() and
removeListener(), so I don’t need to rewrite every bit of listener code.