Fixed test37 trackball freeze on touch screens
Because touch events can contain multiple touches, it’s unwise to have
one layer claim to have handled touchEnd, because it’s critical that
lower layers not miss another touchEnd change event, since those are
used to release the trackball so it can be picked up by another touchId.
That’s what I get for trying to use real events instead of converting
them to something else as soon as they come in.