game2/edit route controls through Player class
Now a Player holds the controls, lays out the on-screen button widgets
& keyboard tips, and passes control impulses to the Spirits under its
control.
Since the controls and the friction timeouts are on different
schedules, I improved the acceleration math to take world-time-passed
into account, so changing the delivery schedule or world clock speed
shouldn’t change the in-world accel (much).
Still hard-coded for one player though.