game2 pause menu for touch
Double-tap the middle of the screen to pause.
The first tap reveals the animated pause-hint. The second (rapid) tap
reveals the pause-menu.
When unpausing by any means, the hint appears in place of the menu and
fades. I’m trying to hint that the center is tappable. Needs user
testing!
It’s in the middle of the screen because that’s the most out-of-the-way
place for a two-player game.