game2 player aim/fire improvements
Added burst fire so player can tap fire and change aim while firing, or
hold fire and lock aim.
Improved formula that updates aim vector, constraining the vec to a
circle, but shifting it based on the square of the trackball movement,
so small adjustments and large changes are both easier, and jumpy touch
controllers don’t result in jumpy aim when the finger moves by tiny
amounts.