game2 explosions dig craters
exploding ants now create a ring of open space around them, with some
wall bits in the middle where they used to be. This causes ants behind
it to fork around the debris  usually, so the player has to move a lot
more, and the levels are constantly changing (and filling w debris)
Weakened the ants since now the player is firing into corpse-dirt a lot
of the time. Now a close shotgun blast will kill one reliably.
The playee explosion digs a crater now too, without any debris in the
middle.