player trackball controls ang accel, not ang vel
now you don’t slam the brakes when resting a finger on the controls,
and it’s easier to spin the ball like a flywheel, adding accel instead
of trying to keep every swipe at max velocity.
It’s less direct, but more intuitive in a way, and much easier to do
big jumps on touch now.