Moved undo/redo and pause into Editor
Tested in test41, and pause is fixed in test40.
But I need to test and fix pause (and that’s all?) in all Editor
clients. Right now pause is probably broken in most editors, because
there are two pause buttons on top of each other, each toggling pause
state when clicked.