Edit screens now provide spiritConfigs to Editor constructor
This consolidates the menu creation code in the editor, and makes the
spiritConfigs param mandatory again.
Test38 was going to be too much trouble to port, and I’ve just replaced
what it was meant to test with a newer system, so I deleted it.
Test37, the first Editor client, didn’t support “add”, so it gets empty
spiritConfigs.