game4 overhauled ant motion
Now they use proper body rotation, and aI made a bunch of misc changes
so their motion and rotation is more fluid, and they seem to avoid
walls better without doing that unnatural sudden reorientation when
they’re under a lot of stress. I like it. And it scales well to large
and small ants