Spring math thing for perfect pulse-accel landings
I’ll use it to make tractor beams with high acceleration but without
oscillation, because it does the second-order math instead of a dumb
spring with damping. It knows how to hit the brakes before reaching the
target position, to hit zero velocity. And it further prevents
overshoot, because it’s aware the acceleration will be applied as a
pulse-force, so it avoids overshooting when close to target.