game4 player slot continuity v1
Now player slots have hardcoded unique names.
The playscreen can create a gameState object that, for now, contains the lsit of active player slot names.
When the players exit a level, a gameState object is sent up from the screen to the page to the app, and back to the next page, which uses it to restore the screen's list of players.
Later that gamestate might include equipment and stats.