game4 telekenisis!
Now an empty-handed kick will repel nearby objects in the general direction of aim, and grab will gradually attract objects that are far away. It's way more fun than before!

This is accomplished with TractorBullets, which are regular invisible beam-bullets with positive or negative mass.

Still needs effects & sound, and limits on repulsor fire rate/power.