more efficient graphing
It's a little hacky, but now the whole graph data array buffer is passed in to GL as a single array, and there's a big model that is just a bunch of line segments with vertex Z vals that indicate which point that vertex belongs to, relative to the head of the circular array. The vertex shader moves each poly-line vertex based on the data array. So far it seems a lot faster. And it should let me scale up the number of points per graph much more cheaply.