centipedes sleep when far from players
This is complicated because it enforces this rule: Either the whole centipede is sleeping, or none of it is.
this.sleepy is a read-often, write-seldom field that is always set to the same value for every spirit in the chain. Without it, each spirit would need to traverse the chain frequently to see if it should wake.