PointerLocker
Helps manages pointer lock based on whether the app thinks it's useful at the moment, and whether the user indicated they want pointerlock or not.
Clicking the page while the pointer lock is useful makes it wanted.
Escaping out of the pointerlock when it's useful and wanted makes it unwanted.