basic targetting scanner
Changed BaseWeapon to BaseTool because the scanner is a lot like a weapon in terms of timeout management and on/off switching, so now they can share the same baseclass.
Added a super primitive scanner that the ants use to see if the player was in their narrow field of vision recently, and hooked that up the laser. It works.