game5: dropping items, game5db and game5key
Players can now drop items, and drop them all when they die (scattering/throwing TBD)
To do that, a PlayerSpirit with an Item needs to know what ItemSpirit to create.
To do that, I added central key ID for in-game objects, that can be used as a spirit type, and can be used to look up spirit constructor, item constructor, editor menu item ID, and modelId, and then I used that everywhere in game5.