lots of item debugging, updated all items
The onTimeout override in BaseTool was ignoring the BaseSpirit timeout, breaking the active and passive timeout loops.
Dropping wasn't wielding the next item.
Laser drawing depended on having a wielder from which to get an aimvec.
Item dropping was happening in die() but it needs to be in explode().
Item buttons stayed down when they were dropped.
Got laser and slow shooter working.
Got ant working again.
Deleted old ItemSpirits.