mine behavior and explosions and fixes
Now mines don't explode in the editor. But they do if you're playing and you touch them with anything, even thrown items. Adjusted the explosion splashes to be less, and the actual exploding projectiles to be more, and added an initial terrain-dig so that mines always leave a good dent.