work around play-time packed-mine perf problem
A dense pack of mines will micro-collide off of each other without doing any damage, so it'll go on for a long time, grinding physics to a halt.
To work around that, make mines damage things a little on contact, so micro-collisions result in death.
But this is a general problem with continuous collision detection that's not totally resolved, just avoided.