Obsoleted ObjSet, and fixed JsonStorage
ObjsSet was only used once, in storage.

JsonStorage wasn't casting things to String on the way in, but it was stringifying on the way out, which is wrong, because it was getting a raw JSON object when it was initialized, which was also wrong. All StorageLikes should store String:String mappings only. (JsonStorage predates native JS Map objects.)