Nuked most flying!
Driving is the best. Flying is OK sometimes but transformations are out, and the only button is the fire button. You can just force your ship off the wall now and hop around the level like crazy. The gun has recoil so you can fly with that, but I might go back to a weak-thrust mode when you're off the wall. This is good though.